convert docs component use a container for data loading
